requirements.txt
boto3
faiss-cpu
pillow
packaging
langchain_community
python-pptx
pymupdf
langchain_openai